ci(prcheck): split pipeline script into client + thin runner
Refactor the prcheck pipeline script in preparation for a second runner
that submits package builds. No behavior change.

* Move .github/workflows/scripts/control-tower-prcheck/ to
  .github/workflows/scripts/control-tower/.
* Extract HTTP/auth/polling helpers from run_control_tower_prcheck.py
  into a shared client.py module (post_scenario, get_job_status,
  poll_until_terminal, etc.).
* Rename the runner to run_prcheck.py and reduce it to a thin CLI on
  top of client.py.
* Update the sources-upload stages template to point at the new
  script paths.
